edge_transport/mqtt_transfer/requirements.txt
paho-mqtt
Pillow